fix: ocr 요소 추출 및 상태값 변경 완료
