+++ package.json
@@ -27,7 +27,8 @@
     "react-native-safe-area-context": "^4.10.1",
     "react-native-screens": "^3.31.1",
     "react-native-web": "~0.19.10",
-    "expo-constants": "~16.0.2"
+    "expo-constants": "~16.0.2",
+    "typescript": "~5.3.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
