+++ src/screens/main.js
@@ -3,7 +3,6 @@ import {
   StyleSheet,
   Text,
   View,
-  TouchableOpacity,
   ScrollView,
   Image,
   TouchableHighlight,
@@ -11,12 +10,12 @@ import {
 import * as ImagePicker from "expo-image-picker";
 import React, { useEffect, useState } from "react";
 import Gifty from "../components/gifty.js";
-import { checkOCR } from "../ocr.js";
-// import * as checkOCR from "../ocr.js";
+import { checkOcr } from "../ocr/checkOcr.js";
 
 const items = [
   { id: 1, image: require("../assets/gif2.png"), pdName: "sef", sName: "df", dd: 33 },
   { id: 2, image: require("../assets/gif2.png"), pdName: "22", sName: "44", dd: 33 },
+
 ];
 
 export default function Main({ navigation }) {
@@ -28,15 +27,12 @@ export default function Main({ navigation }) {
     });
     // console.log("Image Picker Result: ", result);
 
-    let imageUri;
     if (result.assets && result.assets.length > 0 && !result.canceled) {
-      imageUri = result.assets[0].uri;
-      console.log("Selected image URI: ", imageUri);
-      navigation.navigate("Regist", { image: imageUri });
+      const imageUri = result.assets[0].uri;
+      // console.log("Selected image URI: ", imageUri);
+      const giftyconInfo = await checkOcr(imageUri)
+      navigation.navigate("Regist", { image: imageUri, giftyconInfo:giftyconInfo });
     }
-
-    console.log("imageUri", imageUri);
-    checkOCR(imageUri);
   };
 
   return (
